VehicleDemo - Work on the terrain, meshes were a total mess and caused physics problems and horrible frame rate. - Started a basic logic system using nodes
If you drive carefully you can already complete some laps, looping is tough but possible.

Hopefully more work on the physics will come soon but I still need to work on the artwork:
Now the whole trak is one big mesh and cannot recieve shadows anynore. The error says to increase shadow index but maybe it is because the mesh is too big or have too much textures/UVMaps and this can be solved by cleaning the .blend.
It also needs some proper camera control.