Added a sub project related to file loading. It uses a different method of loading files than bParse.
Instead of recursively testing file to memory structures for 
differences it builds an offset table.

This table is used to add file data to memory data in one pass.
Later the offset tables could be saved in the DNA file,  
then the only potentially long lookup times will be when linking 
the two tables together.

It also includes a file compiler, which can be used with a custom CMake
target during build time, so there is no need for separate 
32 / 64 bit DNA files anymore.

The file inspector application uses this new interface to load / save
projects.

This has turned into a very lengthy project, but it is now in in a 
semi - usable state.